+++ src/Farmer/Common.fs
@@ -865,6 +865,34 @@ module Vm =
             | Regular -> "Regular"
             | Spot _ -> "Spot"
 
+module Image =
+    type Architecture =
+        | Arm64
+        | X64
+
+        member this.ArmValue =
+            match this with
+            | Arm64 -> "Arm64"
+            | X64 -> "x64"
+
+    type OsState =
+        | Generalized
+        | Specialized
+
+        member this.ArmValue =
+            match this with
+            | Generalized -> "Generalized"
+            | Specialized -> "Specialized"
+
+    type HyperVGeneration =
+        | V1
+        | V2
+
+        member this.ArmValue =
+            match this with
+            | V1 -> "V1"
+            | V2 -> "V2"
+
 module internal Validation =
     // ANDs two validation rules
     let (<+>) a b v = a v && b v
@@ -1632,7 +1660,6 @@ module GalleryValidation =
         member this.ResourceName =
             match this with
             | GalleryName name -> name
-
 
 module Search =
     type HostingMode =