+++ RELEASE_NOTES.md
@@ -2,8 +2,9 @@ Release Notes
 =============
 
 ## vNext
-* Gallery: Create Galleries for sharing VM images.
-* Image Template: Templates for building and publishing VM images.
+* Galleries: Create Galleries for sharing VM images.
+* Images: Create images in a gallery that can be used to create VMs.
+* Image Templates: Templates for building and publishing VM images.
 
 ## 1.7.17
 * Container Apps: Fix scaler spelling